Move stats to their own file
