+++ src/app/inventory/store/d2-item-factory.service.ts
@@ -1,18 +1,13 @@
 import {
   DestinyClass,
   DestinyInventoryItemDefinition,
-  DestinyInventoryItemStatDefinition,
   DestinyItemComponent,
   DestinyItemComponentSetOfint64,
   DestinyItemInstanceComponent,
-  DestinyItemInvestmentStatDefinition,
   DestinyItemObjectivesComponent,
   DestinyItemSocketsComponent,
-  DestinyItemStatsComponent,
   DestinyItemTalentGridComponent,
   DestinyObjectiveDefinition,
-  DestinyStat,
-  DestinyStatDefinition,
   DestinyTalentGridDefinition,
   ItemLocation,
   TransferStatuses,
@@ -38,7 +33,6 @@ import { t } from 'app/i18next-t';
 import {
   D2Item,
   DimPerk,
-  DimStat,
   DimObjective,
   DimFlavorObjective,
   DimTalentGrid,
@@ -52,7 +46,6 @@ import {
 import { D2Store } from '../store-types';
 import { InventoryBuckets } from '../inventory-buckets';
 import { D2StoresService } from '../d2-stores.service';
-import { filterPlugs } from '../../d2-loadout-builder/generated-sets/utils';
 import { D2CalculatedSeason, D2CurrentSeason } from './../d2-season-info';
 import { D2SourcesToEvent } from 'data/d2/d2-event-info';
 import D2Seasons from 'data/d2/seasons.json';
@@ -60,6 +53,7 @@ import D2SeasonToSource from 'data/d2/seasonToSource.json';
 import D2Events from 'data/d2/events.json';
 import idx from 'idx';
 import { compareBy, chainComparator } from 'app/comparators';
+import { buildStats } from './stats';
 
 // Maps tierType to tierTypeName in English
 const tiers = ['Unknown', 'Currency', 'Common', 'Uncommon', 'Rare', 'Legendary', 'Exotic'];
@@ -73,37 +67,6 @@ let _idTracker: { [id: string]: number } = {};
 const _moveTouchTimestamps = new Map<string, number>();
 
 const SourceToD2Season = D2SeasonToSource.sources;
-
-const statWhiteList = [
-  4284893193, // Rounds Per Minute
-  2961396640, // Charge Time
-  3614673599, // Blast Radius
-  2523465841, // Velocity
-  4043523819, // Impact
-  1240592695, // Range
-  155624089, // Stability
-  943549884, // Handling
-  4188031367, // Reload Speed
-  1345609583, // Aim Assistance
-  3555269338, // Zoom
-  2715839340, // Recoil Direction
-  3871231066, // Magazine
-  2996146975, // Mobility
-  392767087, // Resilience
-  1943323491, // Recovery
-  447667954, // Draw Time
-  1931675084 // Inventory Size
-  //    1935470627, // Power
-  // there are a few others (even an `undefined` stat)
-];
-
-const statsNoBar = [4284893193, 3871231066, 2961396640, 447667954, 1931675084];
-
-const hiddenStats = [
-  1345609583, // Aim Assistance
-  3555269338, // Zoom
-  2715839340 // Recoil Direction
-];
 
 const resistanceMods = {
   1546607977: DamageType.Kinetic,
@@ -448,26 +411,8 @@ export function makeItem(
   }
 
   try {
-    const statsData = idx(itemComponents, (i) => i.stats.data);
-    if (statsData) {
-      // Instanced stats
-      createdItem.stats = buildStats(item, createdItem.sockets, statsData, defs.Stat);
-      if (itemDef.stats && itemDef.stats.stats) {
-        // Hidden stats
-        createdItem.stats = (createdItem.stats || []).concat(
-          buildDefinitionStats(itemDef, defs.Stat, hiddenStats)
-        );
-      }
-    } else if (itemDef.stats && itemDef.stats.stats) {
-      // Item definition stats
-      createdItem.stats = buildDefinitionStats(itemDef, defs.Stat, statWhiteList);
-    }
-    // Investment stats
-    if (!createdItem.stats && itemDef.investmentStats && itemDef.investmentStats.length) {
-      createdItem.stats = buildInvestmentStats(itemDef.investmentStats, defs.Stat);
-    }
-
-    createdItem.stats = createdItem.stats && createdItem.stats.sort(compareBy((s) => s.sort));
+    // See stats.ts
+    createdItem.stats = buildStats(createdItem, itemDef, defs, item, itemComponents);
   } catch (e) {
     console.error(`Error building stats for ${createdItem.name}`, item, itemDef, e);
     reportException('Stats', e, { itemHash: item.itemHash });
@@ -646,184 +591,6 @@ function getSeason(item: D2Item): number {
   }
 
   return D2Seasons[item.hash] || D2CalculatedSeason || D2CurrentSeason;
-}
-
-/**
- * Stats that come from the base definition - we have no info about how they apply to an instance.
- */
-function buildDefinitionStats(
-  itemDef: DestinyInventoryItemDefinition,
-  statDefs: LazyDefinition<DestinyStatDefinition>,
-  statWhiteList: number[]
-): DimStat[] {
-  const itemStats = itemDef.stats.stats;
-
-  if (!itemStats) {
-    return [];
-  }
-
-  return _.compact(
-    _.map(itemStats, (stat: DestinyInventoryItemStatDefinition): DimStat | undefined => {
-      const def = statDefs.get(stat.statHash);
-
-      if (!statWhiteList.includes(stat.statHash) || !stat.value) {
-        return undefined;
-      }
-
-      return {
-        base: stat.value,
-        bonus: 0,
-        statHash: stat.statHash,
-        name: def.displayProperties.name,
-        id: stat.statHash,
-        sort: statWhiteList.indexOf(stat.statHash),
-        value: stat.value,
-        // Armor stats max out at 5, all others are... probably 100? See https://github.com/Bungie-net/api/issues/448
-        maximumValue: [1943323491, 392767087, 2996146975].includes(stat.statHash) ? 5 : 100,
-        bar: !statsNoBar.includes(stat.statHash)
-      };
-    })
-  );
-}
-
-/**
- * Instanced, precalculated stats.
- */
-function buildStats(
-  item: DestinyItemComponent,
-  sockets: DimSockets | null,
-  stats: { [key: string]: DestinyItemStatsComponent },
-  statDefs: LazyDefinition<DestinyStatDefinition>
-): DimStat[] {
-  if (!item.itemInstanceId || !stats[item.itemInstanceId]) {
-    return [];
-  }
-  const itemStats = stats[item.itemInstanceId].stats;
-
-  // determine bonuses for armor
-  const statBonuses = {};
-  if (sockets) {
-    const bonusPerk = sockets.sockets.find((socket) =>
-      Boolean(
-        // Mobility, Restorative, and Resilience perk
-        idx(socket.plug, (p) => p.plugItem.plug.plugCategoryHash) === 3313201758
-      )
-    );
-    // If we didn't find one, then it's not armor.
-    if (bonusPerk) {
-      statBonuses[bonusPerk.plug!.plugItem.investmentStats[0].statTypeHash] = {
-        plugs: bonusPerk.plug!.plugItem.investmentStats[0].value,
-        perks: 0,
-        mods: 0
-      };
-
-      // Custom applied mods
-      sockets.sockets
-        .filter((socket) =>
-          Boolean(
-            idx(socket.plug, (p) => p.plugItem.plug.plugCategoryHash) === 3347429529 &&
-              idx(socket.plug, (p) => p.plugItem.inventory.tierType) !== 2
-          )
-        )
-        .forEach((socket) => {
-          const bonusPerkStat = socket.plug!.plugItem.investmentStats[0];
-          if (bonusPerkStat) {
-            if (!statBonuses[bonusPerkStat.statTypeHash]) {
-              statBonuses[bonusPerkStat.statTypeHash] = { mods: 0 };
-            }
-            statBonuses[bonusPerkStat.statTypeHash].mods += bonusPerkStat.value;
-          }
-        });
-
-      // Look for perks that modify stats (ie. Traction 1818103563)
-      sockets.sockets
-        .filter((socket) =>
-          Boolean(
-            filterPlugs(socket) &&
-              idx(socket.plug, (p) => p.plugItem.plug.plugCategoryHash) !== 3347429529 &&
-              idx(socket.plug, (p) => p.plugItem.investmentStats.length)
-          )
-        )
-        .forEach((socket) => {
-          const bonusPerkStat = socket.plug!.plugItem.investmentStats[0];
-          if (bonusPerkStat) {
-            if (!statBonuses[bonusPerkStat.statTypeHash]) {
-              statBonuses[bonusPerkStat.statTypeHash] = { perks: 0 };
-            }
-            statBonuses[bonusPerkStat.statTypeHash].perks += bonusPerkStat.value;
-          }
-        });
-    }
-  }
-
-  return _.compact(
-    Object.values(itemStats).map((stat: DestinyStat): DimStat | undefined => {
-      const def = statDefs.get(stat.statHash);
-      const itemStat = itemStats[stat.statHash];
-      if (!def || !itemStat) {
-        return undefined;
-      }
-
-      const value = itemStat ? itemStat.value : stat.value;
-      let base = value;
-      let bonus = 0;
-      let plugBonus = 0;
-      let modsBonus = 0;
-      let perkBonus = 0;
-      if (statBonuses[stat.statHash]) {
-        plugBonus = statBonuses[stat.statHash].plugs || 0;
-        modsBonus = statBonuses[stat.statHash].mods || 0;
-        perkBonus = statBonuses[stat.statHash].perks || 0;
-        bonus = plugBonus + perkBonus + modsBonus;
-        base -= bonus;
-      }
-
-      return {
-        base,
-        bonus,
-        plugBonus,
-        modsBonus,
-        perkBonus,
-        statHash: stat.statHash,
-        name: def.displayProperties.name,
-        id: stat.statHash,
-        sort: statWhiteList.indexOf(stat.statHash),
-        value,
-        maximumValue: itemStat.maximumValue,
-        bar: !statsNoBar.includes(stat.statHash)
-      };
-    })
-  );
-}
-
-/**
- * Build stats from the non-pre-sized investment stats.
- */
-function buildInvestmentStats(
-  itemStats: DestinyItemInvestmentStatDefinition[],
-  statDefs: LazyDefinition<DestinyStatDefinition>
-): DimStat[] {
-  return _.compact(
-    _.map(itemStats, (itemStat): DimStat | undefined => {
-      const def = statDefs.get(itemStat.statTypeHash);
-      /* 1935470627 = Power */
-      if (!def || !itemStat || itemStat.statTypeHash === 1935470627) {
-        return undefined;
-      }
-
-      return {
-        base: itemStat.value,
-        bonus: 0,
-        statHash: itemStat.statTypeHash,
-        name: def.displayProperties.name,
-        id: itemStat.statTypeHash,
-        sort: statWhiteList.indexOf(itemStat.statTypeHash),
-        value: itemStat.value,
-        maximumValue: 0,
-        bar: !statsNoBar.includes(itemStat.statTypeHash)
-      };
-    })
-  );
 }
 
 function buildObjectives(
